requirements.txt
boto3==1.26.37
Flask==2.2.2
matplotlib==3.6.2
numpy==1.23.5
opencv_python==4.6.0.66
pandas==1.5.2
Pillow==9.3.0
PyYAML==6.0
requests==2.28.1
scikit_learn==1.2.0
sentence_transformers==2.2.2
tensorflow==2.11.0
torch==1.13.1
